+++ DICOM/Network/IDicomCStoreProvider.cs
@@ -6,5 +6,6 @@ using System.Text;
 namespace Dicom.Network {
 	public interface IDicomCStoreProvider {
 		DicomCStoreResponse OnCStoreRequest(DicomCStoreRequest request);
+		void OnCStoreRequestException(string tempFileName, Exception e);
 	}
 }